Fix the strip row offset wrapping on frames past 4 GiB
The buffer size is computed in usize, but the per-strip offset multiplied
row by stride in u32 before widening. An animation frame whose buffer
exceeds 4 GiB wraps the offset in release and overwrites earlier rows;
the single-frame path is unaffected because it downscales first.

+++ src/image/decode.rs
@@ -2970,8 +2970,9 @@ fn copy_pixels_into(
             Width: width as i32,
             Height: rows as i32,
         };
-        let start = (row * stride) as usize;
-        let end = start + (rows * stride) as usize;
+        // In usize: an animation frame past 4 GiB would wrap a u32 row offset.
+        let start = row as usize * stride as usize;
+        let end = start + rows as usize * stride as usize;
         unsafe { source.CopyPixels(&raw const rectangle, stride, &mut pixels[start..end])? };
         row += rows;
     }
